+++ src/app/components/searchpost/SearchPost.tsx
@@ -1,27 +1,12 @@
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
 import { Post } from '../../types/types'
 import { Link } from 'react-router-dom'
-import { useDispatch, useSelector } from 'react-redux'
-import { RootState } from '../../store/store'
-import { fetchAllPosts } from '../../store/slices/slice'
+
 import './searchpost.css'
-import { getPosts } from '../../services/postService'
 
 function SearchPost() {
   const [query, setQuery] = useState('')
   const [filtered, setFiltered] = useState<Post[]>([])
-  const { posts } = useSelector((state: RootState) => state.posts_reducer)
-  const dispatch = useDispatch()
-
-  useEffect(() => {
-    if (posts.length === 0) {
-      const fetchPosts = async () => {
-        const postsData = await getPosts()
-        dispatch(fetchAllPosts(postsData))
-      }
-      fetchPosts()
-    }
-  }, [dispatch, posts.length])
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value
