+++ src/app/components/subheader/subheader.css
@@ -18,3 +18,9 @@
   margin-left: auto;
   margin-right: 10%;
 }
+
+.subheader__item--language {
+  display: flex;
+  flex-direction: column;
+  margin-left: 2vh;
+}
