+++ src/main.tsx
@@ -4,11 +4,15 @@ import App from './app/App'
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
 import { store } from './app/store/store'
+import { I18nextProvider } from 'react-i18next'
+import i18n from './i18n/config'
 
 createRoot(document.getElementById('root')!).render(
   <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <I18nextProvider i18n={i18n}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </I18nextProvider>
   </Provider>
 )
